+++ boussole/conf/yaml_backend.py
@@ -6,8 +6,9 @@ YAML settings backend
 import yaml
 import pyaml
 
-from boussole.exceptions import SettingsBackendError
-from boussole.conf.base_backend import SettingsBackendBase
+from ..exceptions import SettingsBackendError
+
+from .base_backend import SettingsBackendBase
 
 
 class SettingsBackendYaml(SettingsBackendBase):